There is no need to separate namespace Parry from Parry.combat as everything is combat-related.
