+++ Parry/Character.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using Parry.Combat;
 
 namespace Parry
 {
